Do not add word of the day 2x in dev mode
In dev mode useEffect with empty dependency array runs twice. This
causes word of the day to be added to the "difficult" deck twice. This
bug shouldn't happen in production so it's not a big deal but it was
annoying me during dev mode so I used a simple strategy to work around
this. I simply don't add word to local storage if it already exists in
the deck.